dg: check file existence before deleting

+++ modules/dansguardian/controller/content-filter/dansguardian/init.rb
@@ -23,7 +23,7 @@ class OnBoard
       dg = ContentFilter::DG.new
       #dg.get_status
       if params['delete_all'] == 'on'
-        FileUtils.rm_r dg.root
+        FileUtils.rm_r dg.root if File.exists? dg.root
       end
       if params['initialize_config_files']
         msg = dg.write_all
